requirements.txt
asn1crypto==0.22.0
attrs==17.2.0
Automat==0.6.0
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.0.3
cssselect==1.0.1
enum34==1.1.6
hyperlink==17.3.0
idna==2.5
incremental==17.5.0
ipaddress==1.0.18
lxml==3.8.0
mysql==0.0.1
MySQL-python==1.2.5
parsel==1.2.0
pyasn1==0.3.2
pyasn1-modules==0.0.11
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==17.2.0
queuelib==1.4.2
requests==2.18.3
Scrapy==1.4.0
service-identity==17.0.0
six==1.10.0
Twisted==17.5.0
urllib3==1.22
w3lib==1.18.0
zope.interface==4.4.2
